Use yarn and jest
