+++ .npmrc
@@ -1,0 +1,1 @@
+scripts-prepend-node-path=true
