+++ package.json
@@ -9,6 +9,7 @@
   "main": "index.ts",
   "dependencies": {
     "@types/express": "4.17.2",
+    "@types/supertest": "^2.0.8",
     "body-parser": "1.19.0",
     "dotenv": "8.2.0",
     "express": "4.17.1",
@@ -22,33 +23,33 @@
     "typescript-logging": "0.6.4"
   },
   "devDependencies": {
-    "@types/chai": "4.2.7",
-    "@types/chai-http": "^4.2.0",
-    "@types/mocha": "5.2.7",
-    "chai": "4.2.0",
-    "chai-http": "^4.3.0",
+    "@babel/core": "^7.7.7",
+    "@babel/preset-env": "^7.7.7",
+    "@babel/preset-typescript": "^7.7.7",
+    "@types/jest": "24.0.0",
+    "babel-cli": "^6.26.0",
+    "babel-jest": "^24.9.0",
+    "babel-preset-env": "^1.7.0",
     "concurrently": "3.5.0",
-    "mocha": "6.2.2",
+    "coveralls": "^3.0.9",
+    "jest": "^24.9.0",
+    "superagent": "^5.1.3",
+    "supertest": "^4.0.2",
     "ts-node": "^8.5.4",
     "tslint": "5.18.0",
-    "typescript": "3.7.4",
-    "nyc": "15.0.0",
-    "coveralls": "3.0.2"
+    "typescript": "3.7.4"
   },
   "scripts": {
-    "lint": "npm run lint:src && npm run lint:test",
-    "lint:src": "tslint -c tslint.json 'src/**/*.ts'",
-    "lint:test": "tslint -c tslint.json 'spec/**/*.ts'",
-    "dev": "npm run build && concurrently --kill-others \"tsc -w -p src\" \"nodemon dist/index.js -e js\" || true",
-    "debug": "npm run build && concurrently --kill-others \"tsc -w -p src\" \"nodemon --inspect dist/index.js -e js\"",
-    "test:watch": "mocha -r ts-node/register -w ./dist/spec/**/*.spec.js --watch-extensions js",
-    "test": "npm run lint && npm run build:test && npm run spec-with-coverage",
-    "spec-with-coverage": "NODE_ENV=test nyc mocha -r ts-node/register -w ./dist/spec/**/*.spec.js",
+    "dev": "yarn build && concurrently --kill-others \"tsc -w -p src\" \"nodemon dist/index.js -e js\" || true",
+    "debug": "yarn build && concurrently --kill-others \"tsc -w -p src\" \"nodemon --inspect dist/index.js -e js\" || true",
     "build": "tsc -p src",
-    "build:test": "tsc -p spec",
     "start": "node dist/index.js",
+    "test": "jest --coverage",
+    "lint:src": "tslint -c tslint.json 'src/**/*.ts'",
+    "lint:test": "tslint -c tslint.json 'test/**/*.ts'",
+    "lint": "yarn lint:src && yarn lint:test",
     "clean": "rimraf dist",
-    "coverage": "nyc report --reporter=text-lcov | coveralls"
+    "coveralls": "jest --coverage && cat ./tests/coverage/lcov.info | coveralls"
   },
   "repository": {
     "type": "git",
@@ -59,13 +60,5 @@
   "bugs": {
     "url": "https://fiuba-laboral-v2.atlassian.net"
   },
-  "homepage": "https://github.com/fiuba-laboral-v2/back-end#readme",
-  "nyc": {
-    "exclude": [
-      "coverage/**",
-      "spec/**",
-      "spec{,-*}.js",
-      "**/*.spec.js"
-    ]
-  }
+  "homepage": "https://github.com/fiuba-laboral-v2/back-end#readme"
 }
