+++ test/controllers/root.test.ts
@@ -1,0 +1,10 @@
+import request from "supertest";
+import app from "../../src/app";
+import {OK} from "http-status-codes";
+
+describe("Root path", () => {
+  test("It should give an OK status", async () => {
+    const response = await request(app).get("/");
+    expect(response.status).toEqual(OK);
+  });
+});
