+++ test/tsconfig.json
@@ -1,8 +1,8 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
-    "outDir": "../dist",
-    "baseUrl": "",
+    "outDir": "../dist/test",
+    "baseUrl": "tsconfig",
     "module": "commonjs"
   }
 }
